tests: added unit test cases for all files.

+++ apis/task-controller.go
@@ -5,6 +5,7 @@ import (
 	"TaskSvc/internals/models"
 	"TaskSvc/internals/services"
 	"net/http"
+	"strings"
 
 	"github.com/gin-gonic/gin"
 )
@@ -23,11 +24,19 @@ func (t *TaskController) GetTasks(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, commons.ApiErrorResponse("Failed to fetch tasks", nil))
 		return
 	}
-	c.JSON(http.StatusOK, tasks)
+	c.JSON(http.StatusOK, map[string]interface{}{
+		"total": len(tasks),
+		"tasks": tasks,
+	})
 }
 
 func (t *TaskController) GetTaskById(c *gin.Context) {
 	taskId := c.Param("id")
+	if len(strings.TrimSpace(taskId)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Task ID is required", nil))
+		return
+	}
+
 	task, err := t.taskService.GetTaskById(c, taskId)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, commons.ApiErrorResponse("Failed to fetch task", nil))
@@ -37,13 +46,27 @@ func (t *TaskController) GetTaskById(c *gin.Context) {
 }
 
 func (t *TaskController) CreateTask(c *gin.Context) {
-	var task models.Task
-	if err := c.ShouldBindJSON(&task); err != nil {
+	var task *models.Task
+	if err := c.ShouldBindJSON(&task); err != nil || task == nil {
 		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Invalid request payload", nil))
 		return
 	}
 
-	taskId, err := t.taskService.CreateTask(c, &task)
+	if len(strings.TrimSpace(task.Title)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Title is required", nil))
+		return
+	}
+
+	if len(strings.TrimSpace(task.Description)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Description is required", nil))
+		return
+	}
+
+	if len(strings.TrimSpace(task.Status)) == 0 {
+		task.Status = "Pending" // Default status if not provided
+	}
+
+	taskId, err := t.taskService.CreateTask(c, task)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, commons.ApiErrorResponse("Failed to create task", nil))
 		return
@@ -53,13 +76,28 @@ func (t *TaskController) CreateTask(c *gin.Context) {
 
 func (t *TaskController) UpdateTask(c *gin.Context) {
 	taskId := c.Param("id")
-	var task models.Task
+	if len(strings.TrimSpace(taskId)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Task ID is required", nil))
+		return
+	}
+
+	var task *models.Task
 	if err := c.ShouldBindJSON(&task); err != nil {
 		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Invalid request payload", nil))
 		return
 	}
 
-	if err := t.taskService.UpdateTask(c, &task, taskId); err != nil {
+	if len(strings.TrimSpace(task.Title)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Title is required", nil))
+		return
+	}
+
+	if len(strings.TrimSpace(task.Description)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Description is required", nil))
+		return
+	}
+
+	if err := t.taskService.UpdateTask(c, task, taskId); err != nil {
 		c.JSON(http.StatusInternalServerError, commons.ApiErrorResponse("Failed to update task", nil))
 		return
 	}
@@ -68,9 +106,15 @@ func (t *TaskController) UpdateTask(c *gin.Context) {
 
 func (t *TaskController) DeleteTask(c *gin.Context) {
 	taskId := c.Param("id")
+	if len(strings.TrimSpace(taskId)) == 0 {
+		c.JSON(http.StatusBadRequest, commons.ApiErrorResponse("Task ID is required", nil))
+		return
+	}
+
 	if err := t.taskService.DeleteTaskById(c, taskId); err != nil {
 		c.JSON(http.StatusInternalServerError, commons.ApiErrorResponse("Failed to delete task", nil))
 		return
 	}
+
 	c.Status(http.StatusNoContent)
 }
